% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-package.R
\docType{package}
\name{useicd10cmShiny-package}
\alias{useicd10cmShiny}
\alias{useicd10cmShiny-package}
\title{useicd10cmShiny: What the Package Does (One Line, Title Case)}
\description{
A shiny app companion of the package useicd10cm.
}
\keyword{internal}
